PlantDCore Cleanup (#51)
* fix: logging in dataset controller

* refactor: rename and extract util funcs

* fix: metrics svc must be in pipeline ns and pipeline label finalizer

* feat: finalizer in pipeline and experiment, experiment draining time

* fix: various issues in controllers

* refactor: plantdcore crd and controller

* fix: various issues

* fix: plantdcore not updating and showing status

* fix: use statefulset to persist redis data

* docs: update comment

+++ api/v1alpha1/loadpattern_types.go
@@ -10,7 +10,7 @@ type Stage struct {
 	// Equivalent to the "ramping-arrival-rate" executor's `stages[].target` option in K6.
 	// See https://k6.io/docs/using-k6/scenarios/executors/ramping-arrival-rate/#options for more details.
 	// +kubebuilder:validation:Minimum=0
-	Target int `json:"target"`
+	Target int64 `json:"target"`
 	// Duration of the stage, also the time to reach the target load.
 	// Equivalent to the "ramping-arrival-rate" executor's `stages[].duration` option in K6.
 	// See https://k6.io/docs/using-k6/scenarios/executors/ramping-arrival-rate/#options for more details.
@@ -22,16 +22,18 @@ type LoadPatternSpec struct {
 	// List of stages in the LoadPattern.
 	// Equivalent to the "ramping-arrival-rate" executor's `stages` option in K6.
 	// See https://k6.io/docs/using-k6/scenarios/executors/ramping-arrival-rate/#options for more details.
+	// +kubebuilder:validation:MinItems=1
 	Stages []Stage `json:"stages"`
 	// Number of VUs to pre-allocate before Experiment start.
 	// Equivalent to the "ramping-arrival-rate" executor's `preAllocatedVUs` option in K6.
 	// See https://k6.io/docs/using-k6/scenarios/executors/ramping-arrival-rate/#options for more details.
 	// +kubebuilder:validation:Minimum=0
-	PreAllocatedVUs int `json:"preAllocatedVUs,omitempty"`
+	PreAllocatedVUs int64 `json:"preAllocatedVUs,omitempty"`
 	// Number of requests per `timeUnit` period at Experiment start.
 	// Equivalent to the "ramping-arrival-rate" executor's `startRate` option in K6.
 	// See https://k6.io/docs/using-k6/scenarios/executors/ramping-arrival-rate/#options for more details.
-	StartRate int `json:"startRate"`
+	// +kubebuilder:validation:Minimum=0
+	StartRate int64 `json:"startRate"`
 	// Period of time to apply to the `startRate` and `stages[].target` fields.
 	// Equivalent to the "ramping-arrival-rate" executor's `timeUnit` option in K6.
 	// See https://k6.io/docs/using-k6/scenarios/executors/ramping-arrival-rate/#options for more details.
@@ -39,7 +41,8 @@ type LoadPatternSpec struct {
 	// Maximum number of VUs to allow for allocation during Experiment.
 	// Equivalent to the "ramping-arrival-rate" executor's `maxVUs` option in K6.
 	// See https://k6.io/docs/using-k6/scenarios/executors/ramping-arrival-rate/#options for more details.
-	MaxVUs int `json:"maxVUs,omitempty"`
+	// +kubebuilder:validation:Minimum=0
+	MaxVUs int64 `json:"maxVUs,omitempty"`
 }
 
 // LoadPatternStatus defines the observed state of LoadPattern.
